requirements.txt
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg==3.0.12
python-dotenv==0.13.0
Werkzeug==2.0.3
